backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
beautifulsoup4==4.12.2
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
lxml==4.9.3
aiohttp==3.9.1
chromadb==0.4.18
